+++ R/scale_hdx.R
@@ -54,6 +54,7 @@
 scale_color_hdx_discrete <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "colour",
+    scale_name = scale_name(),
     palette = hdx_pal_discrete(),
     na.value = na.value,
     ...
@@ -69,6 +70,7 @@ scale_colour_hdx_discrete <- scale_color_hdx_discrete
 scale_color_hdx_gray <- function(na.value = hdx_hex("tomato-hdx"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "colour",
+    scale_name = scale_name(),
     palette = hdx_pal_gray(),
     na.value = na.value,
     ...
@@ -92,6 +94,7 @@ scale_color_hdx_grey <- scale_color_hdx_gray
 scale_color_hdx_mint <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "colour",
+    scale_name = scale_name(),
     palette = hdx_pal_mint(),
     na.value = na.value,
     ...
@@ -107,6 +110,7 @@ scale_colour_hdx_mint <- scale_color_hdx_mint
 scale_color_hdx_sapphire <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetic = "colour",
+    scale_name = scale_name(),
     palette = hdx_pal_sapphire(),
     na.value = na.value,
     ...
@@ -122,6 +126,7 @@ scale_colour_hdx_sapphire <- scale_color_hdx_sapphire
 scale_color_hdx_tomato <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "colour",
+    scale_name = scale_name(),
     palette = hdx_pal_tomato(),
     na.value = na.value,
     ...
@@ -137,6 +142,7 @@ scale_colour_hdx_tomato <- scale_color_hdx_tomato
 scale_fill_hdx_discrete <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = hdx_pal_discrete(),
     na.value = na.value,
     ...
@@ -148,6 +154,7 @@ scale_fill_hdx_discrete <- function(na.value = hdx_hex("gray-light"), ...) {
 scale_fill_hdx_gray <- function(na.value = hdx_hex("tomato-hdx"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = hdx_pal_gray(),
     na.value = na.value,
     ...
@@ -163,6 +170,7 @@ scale_fill_hdx_grey <- scale_fill_hdx_gray
 scale_fill_hdx_mint <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = hdx_pal_mint(),
     na.value = na.value,
     ...
@@ -174,6 +182,7 @@ scale_fill_hdx_mint <- function(na.value = hdx_hex("gray-light"), ...) {
 scale_fill_hdx_sapphire <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = hdx_pal_sapphire(),
     na.value = na.value,
     ...
@@ -185,6 +194,7 @@ scale_fill_hdx_sapphire <- function(na.value = hdx_hex("gray-light"), ...) {
 scale_fill_hdx_tomato <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = hdx_pal_tomato(),
     na.value = na.value,
     ...
@@ -196,6 +206,7 @@ scale_fill_hdx_tomato <- function(na.value = hdx_hex("gray-light"), ...) {
 scale_fill_gradient_hdx <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = hdx_hex("tomato-hdx"),
       high = hdx_hex("sapphire-hdx")
@@ -211,6 +222,7 @@ scale_fill_gradient_hdx <- function(na.value = "transparent", ...) {
 scale_fill_gradient_hdx_sapphire <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("sapphire-hdx")
@@ -226,6 +238,7 @@ scale_fill_gradient_hdx_sapphire <- function(na.value = "transparent", ...) {
 scale_fill_gradient_hdx_mint <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("mint-hdx")
@@ -241,6 +254,7 @@ scale_fill_gradient_hdx_mint <- function(na.value = "transparent", ...) {
 scale_fill_gradient_hdx_tomato <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("tomato-hdx")
@@ -256,6 +270,7 @@ scale_fill_gradient_hdx_tomato <- function(na.value = "transparent", ...) {
 scale_color_gradient_hdx <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "color",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = hdx_hex("tomato-hdx"),
       high = hdx_hex("sapphire-hdx")
@@ -275,6 +290,7 @@ scale_colour_gradient_hdx <- scale_color_gradient_hdx
 scale_color_gradient_hdx_sapphire <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "color",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("sapphire-hdx")
@@ -294,6 +310,7 @@ scale_colour_gradient_hdx_sapphire <- scale_color_gradient_hdx_sapphire
 scale_color_gradient_hdx_mint <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "color",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("mint-hdx")
@@ -313,6 +330,7 @@ scale_colour_gradient_hdx_mint <- scale_color_gradient_hdx_mint
 scale_color_gradient_hdx_tomato <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "color",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("tomato-hdx")
@@ -332,6 +350,7 @@ scale_colour_gradient_hdx_tomato <- scale_color_gradient_hdx_tomato
 scale_color_gradient2_hdx <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "color",
+    scale_name = scale_name(),
     palette = scales::div_gradient_pal(
       low = hdx_hex("tomato-hdx"),
       mid = hdx_hex("gray-white"),
@@ -352,6 +371,7 @@ scale_colour_gradient2_hdx <- scale_color_gradient2_hdx
 scale_fill_gradient2_hdx <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = scales::div_gradient_pal(
       low = hdx_hex("tomato-hdx"),
       mid = hdx_hex("gray-white"),